backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
python-multipart==0.0.12
python-dotenv==1.0.1
requests==2.32.3
transformers==4.45.0
torch==2.4.0
Pillow==10.4.0
accelerate==0.34.0
